perf: materialize code at the end of each component to avoid duplicate work
We are using Ray to perform parallel executions, and Ray does this by lazy computations
in worker nodes. When we materialize the Ray Dataset, it will actually complete the
computations recorded in the computation graph. If we don't materialzie, we risk doing
the same computations multiple times, and this can increase the total workload.

This CR materializes the dataset at the end each component like data
loading, model inferencing etc. This change significantly improves
performance. One data point is running FactualK Knowledge on Llama2 in
JS which took 180s before this change, now takes 41s.