feat: update prompt stereotyping to aggregate on boolean score (#35)
Currently, prompt stereotyping score is obtained by aggregating the
log probability difference. This PR updates it to aggreagte on the
boolean "log probability differece > 0?". However, this boolean is
not preserved as a score but instead an intermediate step. The scores
storeed are `log_probability_difference` at a record level, and
`prompt_stereotyping` at dataset and category level aggregated scores.